fix: :adhesive_bandage: fix sonar cloud code coverage
